refactor(gradle): move `gradleJvmArg` to `util/exec`
As this is the better place for this to be overall, and as well as
future changes that will need to pull it from a central location.
