+++ lib/modules/manager/gradle-wrapper/artifacts.ts
@@ -2,10 +2,13 @@ import { lang, query as q } from '@renovatebot/good-enough-parser';
 import { isTruthy } from '@sindresorhus/is';
 import { quote } from 'shlex';
 import upath from 'upath';
-import type { ToolSettingsOptions } from '../../../config/types.ts';
 import { TEMPORARY_ERROR } from '../../../constants/error-messages.ts';
 import { logger } from '../../../logger/index.ts';
-import { exec, getToolSettingsOptions } from '../../../util/exec/index.ts';
+import {
+  exec,
+  getToolSettingsOptions,
+  gradleJvmArg,
+} from '../../../util/exec/index.ts';
 import type { ExecOptions } from '../../../util/exec/types.ts';
 import {
   localPathExists,
@@ -130,10 +133,6 @@ export async function updateLockFiles(
     newPackageFileContent: buildFileContent,
     config,
   });
-}
-
-export function gradleJvmArg(config: ToolSettingsOptions): string {
-  return ` -Dorg.gradle.jvmargs="-Xms${config.jvmMemory}m -Xmx${config.jvmMaxMemory}m"`;
 }
 
 export async function updateArtifacts({
